tests: move old tests from top level directory to proper place
