+++ tests/acceptance/features/items/super_editor/edit.feature
@@ -14,7 +14,6 @@ Feature: Editing a item
       | List enum in source   | 1                    |
       | Abbreviated statement | New Abb statement    |
       | Concept keywords      | reading              |
-      | Concept keywords uri  | http://reading.com   |
       | Licence uri           | http://somewhere.com |
 
     Then I should see the Item
